Move stepQuizHint to DefaultStepQuizFragment to load it only after attempt StepQuizFeature loading
