+++ androidHyperskillApp/src/main/java/org/hyperskill/app/android/step_practice/view/fragment/StepPracticeFragment.kt
@@ -2,7 +2,6 @@ package org.hyperskill.app.android.step_practice.view.fragment
 
 import android.os.Bundle
 import android.view.View
-import androidx.core.view.isVisible
 import androidx.fragment.app.Fragment
 import by.kirich1409.viewbindingdelegate.viewBinding
 import org.hyperskill.app.android.R
@@ -14,16 +13,13 @@ import org.hyperskill.app.android.step.view.model.StepCompletionView
 import org.hyperskill.app.android.step.view.screen.StepScreen
 import org.hyperskill.app.android.step_content_text.view.fragment.TextStepContentFragment
 import org.hyperskill.app.android.step_quiz.view.factory.StepQuizFragmentFactory
-import org.hyperskill.app.android.step_quiz_hints.fragment.StepQuizHintsFragment
 import org.hyperskill.app.step.domain.model.Step
 import org.hyperskill.app.step.domain.model.StepRoute
-import org.hyperskill.app.step_quiz_hints.presentation.StepQuizHintsFeature
 
 class StepPracticeFragment : Fragment(R.layout.fragment_step_practice), StepCompletionView {
     companion object {
         private const val STEP_CONTENT_FRAGMENT_TAG = "step_content"
         private const val STEP_QUIZ_FRAGMENT_TAG = "step_quiz"
-        private const val STEP_HINTS_FRAGMENT_TAG = "step_hints"
 
         fun newInstance(step: Step, stepRoute: StepRoute): Fragment =
             StepPracticeFragment().apply {
@@ -55,7 +51,6 @@ class StepPracticeFragment : Fragment(R.layout.fragment_step_practice), StepComp
         }
         initStepTheoryFragment(step)
         setStepQuizFragment(step, stepRoute)
-        setStepHintsFragment(step)
     }
 
     private fun initStepTheoryFragment(step: Step) {
@@ -67,16 +62,6 @@ class StepPracticeFragment : Fragment(R.layout.fragment_step_practice), StepComp
     private fun setStepQuizFragment(step: Step, stepRoute: StepRoute) {
         setChildFragment(R.id.stepQuizContainer, STEP_QUIZ_FRAGMENT_TAG) {
             StepQuizFragmentFactory.getQuizFragment(step, stepRoute)
-        }
-    }
-
-    private fun setStepHintsFragment(step: Step) {
-        val isFeatureEnabled = StepQuizHintsFeature.isHintsFeatureAvailable(step)
-        viewBinding.stepQuizHints.isVisible = isFeatureEnabled
-        if (isFeatureEnabled) {
-            setChildFragment(R.id.stepQuizHints, STEP_HINTS_FRAGMENT_TAG) {
-                StepQuizHintsFragment.newInstance(step)
-            }
         }
     }
