+++ androidHyperskillApp/src/main/res/layout/fragment_stage_step_wrapper.xml
@@ -43,14 +43,6 @@
                 />
 
             <FrameLayout
-                android:id="@+id/stageHints"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingHorizontal="20dp"
-                android:visibility="gone"
-                />
-
-            <FrameLayout
                 android:id="@+id/stageQuizContainer"
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
