+++ androidHyperskillApp/src/main/res/layout/fragment_step_practice.xml
@@ -31,14 +31,6 @@
                 />
 
             <FrameLayout
-                android:id="@+id/stepQuizHints"
-                android:layout_width="match_parent"
-                android:layout_height="wrap_content"
-                android:paddingHorizontal="20dp"
-                android:visibility="gone"
-                />
-
-            <FrameLayout
                 android:id="@+id/stepQuizContainer"
                 android:layout_width="match_parent"
                 android:layout_height="0dp"
