+++ androidHyperskillApp/src/main/res/layout/fragment_step_quiz.xml
@@ -31,6 +31,17 @@
 
         app:layout_constraintVertical_bias="0" />
 
+    <FrameLayout
+        android:id="@+id/stepQuizHints"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        app:layout_constraintTop_toTopOf="parent"
+        app:layout_constraintStart_toStartOf="parent"
+        app:layout_constraintEnd_toEndOf="parent"
+        android:layout_marginHorizontal="20dp"
+        android:visibility="gone"
+        />
+
     <androidx.appcompat.widget.AppCompatTextView
         android:id="@+id/stepQuizStatistics"
         android:layout_width="0dp"
