+++ androidHyperskillApp/src/main/res/layout/layout_step_quiz_choice.xml
@@ -8,7 +8,7 @@
     android:orientation="vertical"
     android:layout_marginVertical="20dp"
     app:layout_constraintBottom_toTopOf="@id/stepQuizFeedbackBlocks"
-    app:layout_constraintTop_toTopOf="parent"
+    app:layout_constraintTop_toBottomOf="@+id/stepQuizHints"
     app:layout_constraintEnd_toEndOf="parent"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintVertical_bias="0">
