+++ androidHyperskillApp/src/main/res/layout/layout_step_quiz_code.xml
@@ -6,7 +6,7 @@
     android:layout_marginVertical="20dp"
     app:layout_constraintStart_toStartOf="parent"
     app:layout_constraintEnd_toEndOf="parent"
-    app:layout_constraintTop_toTopOf="parent"
+    app:layout_constraintTop_toBottomOf="@+id/stepQuizHints"
     app:layout_constraintBottom_toTopOf="@id/stepQuizFeedbackBlocks"
     app:layout_constraintVertical_bias="0" >
