+++ androidHyperskillApp/src/main/res/layout/layout_step_quiz_sorting.xml
@@ -10,7 +10,7 @@
     app:layout_constraintBottom_toTopOf="@id/stepQuizFeedbackBlocks"
     app:layout_constraintEnd_toEndOf="parent"
     app:layout_constraintStart_toStartOf="parent"
-    app:layout_constraintTop_toTopOf="parent">
+    app:layout_constraintTop_toBottomOf="@+id/stepQuizHints" >
 
     <include
         android:id="@+id/sortingStepDescription"
